Implemented user local storage. Room storage with host inside needs testing

+++ app/src/main/java/com/example/roomer/room/entities/LocalRoom.kt
@@ -8,7 +8,7 @@ data class LocalRoom(
     @PrimaryKey
     val roomId: Int,
     val monthPrice: Int,
-    val hostId: Long,
+    val hostId: Int,
     val description: String,
     val photo: String,
     val bathroomsCount: Int,
